Fixing size_t

+++ srcs/CGI/handling_cgi.cpp
@@ -6,7 +6,7 @@
 /*   By: zwina <zwina@student.1337.ma>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/12 12:24:22 by zoukaddo          #+#    #+#             */
-/*   Updated: 2023/08/09 09:10:27 by zwina            ###   ########.fr       */
+/*   Updated: 2023/08/09 09:26:24 by zwina            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -104,7 +104,7 @@ std::string Response::env_grabber(const std::string& key)
 
 void Response::reqbodysend()
 {
-    int content_length;
+    size_t content_length;
     
     std::map<std::string, std::string>::iterator it = request->_fields.find("CONTENT-LENGTH");
     if (it != request->_fields.end())
